requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.3
django-extensions==3.2.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
exceptiongroup==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
simplejson==3.19.3
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
